:recycle: 함수 분리
