+++ app/src/androidTest/java/nextstep/signup/TestFixture.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import nextstep.signup.ui.CustomTextField
 
 object TestFixture {
     const val TEST_TAG = "이름"
